requirements.txt
selenium
pandas
beautifulsoup4
requests
azure.storage.blob